Use fewer chars

+++ util/rand.go
@@ -4,7 +4,7 @@ import "math/rand/v2"
 
 var LettersLower = []rune("abcdefghijklmnopqrstuvwxyz")
 
-func RandomString(n int, charPool []rune) string {
+func RandomString(n uint64, charPool []rune) string {
 	s := make([]rune, n)
 	for i := range s {
 		s[i] = charPool[rand.IntN(len(charPool))]
